Trait oasis_contract_sdk_storage::map::Integer

pub trait Integer {
    type Encoded: AsRef<[u8]>;

    // Required method
    fn to_be_bytes(self) -> Self::Encoded;
}

A trait representing an integer that can be encoded into big-endian bytes.

Required Associated Types

type Encoded: AsRef<[u8]>

Type of the encoded representation.

Required Methods

fn to_be_bytes(self) -> Self::Encoded

Return the memory representation of this integer as a byte array in big-endian byte order.